+++ contracts/azorius/BaseStrategy.sol
@@ -37,13 +37,13 @@ abstract contract BaseStrategy is OwnableUpgradeable, FactoryFriendly, IBaseStra
     function initializeProposal(bytes memory _data) external virtual;
 
     /** @inheritdoc IBaseStrategy*/
-    function isPassed(uint256 _proposalId) external view virtual returns (bool);
+    function isPassed(uint32 _proposalId) external view virtual returns (bool);
 
     /** @inheritdoc IBaseStrategy*/
     function isProposer(address _address) external view virtual returns (bool);
 
     /** @inheritdoc IBaseStrategy*/
-    function votingEndBlock(uint256 _proposalId) external view virtual returns (uint32);
+    function votingEndBlock(uint32 _proposalId) external view virtual returns (uint32);
 
     /**
      * Sets the address of the Azorius module contract.
