+++ contracts/azorius/interfaces/IBaseStrategy.sol
@@ -34,7 +34,7 @@ interface IBaseStrategy {
      * @param _proposalId proposalId to check
      * @return bool true if the proposal has passed, otherwise false
      */
-    function isPassed(uint256 _proposalId) external view returns (bool);
+    function isPassed(uint32 _proposalId) external view returns (bool);
 
     /**
      * Returns whether the specified address can submit a Proposal with
@@ -55,5 +55,5 @@ interface IBaseStrategy {
      * @param _proposalId proposalId to check
      * @return uint32 block number when voting ends on the Proposal
      */
-    function votingEndBlock(uint256 _proposalId) external view returns (uint32);
+    function votingEndBlock(uint32 _proposalId) external view returns (uint32);
 }
